+++ src/config/chain/initia.ts
@@ -2,6 +2,9 @@ import { wallets as keplrWallets } from "@cosmos-kit/keplr";
 
 import type { ChainConfigs } from "./types";
 
+const INITIA_DECODER =
+  "https://initia-api-jiod42ec2q-as.a.run.app/decode_module";
+
 export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
   "mahalo-1": {
     chain: "initia",
@@ -22,7 +25,7 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
       move: {
         enabled: true,
         moduleMaxFileSize: 1_048_576,
-        decodeApi: "https://initia-api-jiod42ec2q-as.a.run.app/decode_module",
+        decodeApi: INITIA_DECODER,
       },
       pool: {
         enabled: false,
@@ -52,6 +55,102 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
     },
     extra: {},
   },
+  "minimove-1": {
+    chain: "initia",
+    registryChainName: "minitiamovedevnet1",
+    prettyName: "Minitia Move Devnet 1",
+    lcd: "https://lcd.minimove-1.initia.xyz",
+    rpc: "https://rpc.minimove-1.initia.xyz:443",
+    indexer: "https://minimove-1-graphql.alleslabs.dev/v1/graphql",
+    wallets: [...keplrWallets],
+    features: {
+      faucet: {
+        enabled: false,
+      },
+      wasm: {
+        enabled: false,
+      },
+      move: {
+        enabled: true,
+        moduleMaxFileSize: 1_048_576,
+        decodeApi: INITIA_DECODER,
+      },
+      pool: {
+        enabled: false,
+      },
+      publicProject: {
+        enabled: true,
+      },
+      gov: {
+        enabled: false,
+      },
+      nft: {
+        enabled: true,
+      },
+    },
+    gas: {
+      gasPrice: {
+        tokenPerGas: 0.151,
+        denom: "umin",
+      },
+      gasAdjustment: 1.5,
+      maxGasLimit: 25_000_000,
+    },
+    explorerLink: {
+      validator:
+        "https://lcd.minimove-1.initia.xyz/opinit/opchild/v1/validator",
+      proposal: "",
+    },
+    extra: { disableDelegation: true },
+  },
+  "miniwasm-1": {
+    chain: "initia",
+    registryChainName: "minitiawasmdevnet1",
+    prettyName: "Minitia Wasm Devnet 1",
+    lcd: "https://lcd.miniwasm-1.initia.xyz",
+    rpc: "https://rpc.miniwasm-1.initia.xyz:443",
+    indexer: "https://miniwasm-1-graphql.alleslabs.dev/v1/graphql",
+    wallets: [...keplrWallets],
+    features: {
+      faucet: {
+        enabled: false,
+      },
+      wasm: {
+        enabled: true,
+        storeCodeMaxFileSize: 800_000,
+        clearAdminGas: 50_000,
+      },
+      move: {
+        enabled: false,
+      },
+      pool: {
+        enabled: false,
+      },
+      publicProject: {
+        enabled: true,
+      },
+      gov: {
+        enabled: false,
+      },
+      nft: {
+        enabled: true,
+      },
+    },
+    gas: {
+      gasPrice: {
+        tokenPerGas: 0.151,
+        denom: "umin",
+      },
+      gasAdjustment: 1.5,
+      maxGasLimit: 25_000_000,
+    },
+    explorerLink: {
+      validator:
+        "https://lcd.miniwasm-1.initia.xyz/opinit/opchild/v1/validator",
+      proposal: "",
+    },
+    extra: { disableDelegation: true },
+  },
   "stone-12-1": {
     chain: "initia",
     registryChainName: "initiatestnet12-1",
@@ -71,7 +170,7 @@ export const INITIA_CHAIN_CONFIGS: ChainConfigs = {
       move: {
         enabled: true,
         moduleMaxFileSize: 1_048_576,
-        decodeApi: "https://initia-api-jiod42ec2q-as.a.run.app/decode_module",
+        decodeApi: INITIA_DECODER,
       },
       pool: {
         enabled: false,