+++ src/lib/pages/contract-details/components/tables/RelatedProposalsTable.tsx
@@ -2,7 +2,7 @@ import type { ChangeEvent } from "react";
 
 import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
-import { EmptyState } from "lib/components/state";
+import { EmptyState, ErrorFetching } from "lib/components/state";
 import { ProposalsTable } from "lib/components/table";
 import { useRelatedProposalsByContractAddressPagination } from "lib/services/proposalService";
 import type { ContractAddr, Option } from "lib/types";
@@ -36,12 +36,15 @@ export const RelatedProposalsTable = ({
     },
   });
 
-  const { data: relatedProposals, isLoading } =
-    useRelatedProposalsByContractAddressPagination(
-      contractAddress,
-      offset,
-      pageSize
-    );
+  const {
+    data: relatedProposals,
+    isLoading,
+    error,
+  } = useRelatedProposalsByContractAddressPagination(
+    contractAddress,
+    offset,
+    pageSize
+  );
 
   const onPageChange = (nextPage: number) => {
     refetchCount();
@@ -61,10 +64,14 @@ export const RelatedProposalsTable = ({
         proposals={relatedProposals}
         isLoading={isLoading}
         emptyState={
-          <EmptyState
-            imageVariant="empty"
-            message="This contract does not have related proposals yet."
-          />
+          error ? (
+            <ErrorFetching dataName="related proposals" />
+          ) : (
+            <EmptyState
+              imageVariant="empty"
+              message="This contract does not have related proposals yet."
+            />
+          )
         }
       />
       {!!totalData && totalData > 10 && (