+++ src/lib/pages/module-details/components/module-fns/index.ts
@@ -1,0 +1,2 @@
+export * from "./FunctionTypeSwitch";
+export * from "./ModuleFunctions";